MarkedIcon fully self-contained, and only receives resourceIds and showCount
