+++ src/components/ResourceCard/ResourceCard.js
@@ -20,13 +20,11 @@ import {
   patientSelector,
   patientAgeAtResourcesSelector,
   collectionResourceIdsSelector,
-  markedResourcesSelector,
 } from '../../redux/selectors';
 import { getResourceDate } from '../../resources/fhirReader';
 import Colors from '../../constants/Colors';
 import CountIcon from '../Icons/CountIcon';
 import MarkedIcon from '../Icons/MarkedIcon';
-import { actionTypes } from '../../redux/epics';
 
 const selectCardBody = (resource, patientAgeAtResource) => {
   switch (resource.type) {
@@ -78,8 +76,6 @@ const ResourceCard = ({
   addResourceToCollection,
   removeResourceFromCollection,
   collectionResourceIds,
-  updateMarkedResources,
-  markedResources,
 }) => {
   const resource = resources[resourceId];
   const resourceType = SINGULAR_RESOURCE_TYPES[resource?.type];
@@ -101,12 +97,6 @@ const ResourceCard = ({
     );
   }
 
-  const handleMarkedClick = () => {
-    updateMarkedResources({
-      [resourceId]: !markedResources.marked[resourceId],
-    });
-  };
-
   return (
     <View style={styles.root}>
       <View style={styles.header}>
@@ -114,8 +104,8 @@ const ResourceCard = ({
         <View style={styles.dateIconContainer}>
           <BaseText>{resourceDate}</BaseText>
           <MarkedIcon
-            onClick={handleMarkedClick}
-            hasMarked={!!markedResources.marked[resourceId]}
+            resourceIds={[resourceId]}
+            showCount={false}
           />
           <CountIcon
             shape="square"
@@ -145,11 +135,6 @@ ResourceCard.propTypes = {
   addResourceToCollection: func.isRequired,
   removeResourceFromCollection: func.isRequired,
   collectionResourceIds: shape({}),
-  updateMarkedResources: func.isRequired,
-  markedResources: shape({
-    marked: shape({}).isRequired,
-    lastMarked: shape({}).isRequired,
-  }).isRequired,
 };
 
 ResourceCard.defaultProps = {
@@ -161,17 +146,9 @@ const mapStateToProps = (state) => ({
   patientResource: patientSelector(state),
   patientAgeAtResources: patientAgeAtResourcesSelector(state),
   collectionResourceIds: collectionResourceIdsSelector(state),
-  markedResources: markedResourcesSelector(state),
 });
 
-const mapDispatchToProps = {
-  updateMarkedResources: (resourceIdsMap) => ({
-    type: actionTypes.UPDATE_MARKED_RESOURCES,
-    payload: resourceIdsMap,
-  }),
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(React.memo(ResourceCard));
+export default connect(mapStateToProps, null)(ResourceCard);
 
 const styles = StyleSheet.create({
   root: {
