+++ src/components/SubTypeAccordion/SubTypeAccordion.js
@@ -6,19 +6,14 @@ import { Accordion } from 'native-base';
 import { connect } from 'react-redux';
 
 import {
-  arrayOf, func, number, shape, string,
+  arrayOf, func, number, string,
 } from 'prop-types';
 import Colors from '../../constants/Colors';
 import ResourceCard from '../ResourceCard/ResourceCard';
-import {
-  actionTypes,
-  addResourceToCollection,
-  removeResourceFromCollection,
-} from '../../redux/epics';
+import { addResourceToCollection, removeResourceFromCollection } from '../../redux/epics';
 import BaseText from '../Generic/BaseText';
 import CountIcon from '../Icons/CountIcon';
 import MarkedIcon from '../Icons/MarkedIcon';
-import { markedResourcesSelector } from '../../redux/selectors';
 
 const SubTypeAccordion = ({
   addResourceToCollectionAction,
@@ -28,23 +23,8 @@ const SubTypeAccordion = ({
   resourceIds,
   selectedCollectionId,
   subType,
-  updateMarkedResources,
-  markedResources,
 }) => {
   const dataArray = [{ title: subType, content: resourceIds }];
-
-  const markedResourceCount = resourceIds.reduce((acc, id) => {
-    const isMarked = markedResources.marked[id];
-    return isMarked ? acc + 1 : acc;
-  }, 0);
-
-  const handleMarkedClick = () => {
-    updateMarkedResources(resourceIds.reduce((acc, id) => ({
-      ...acc,
-      [id]: !markedResourceCount,
-    }), {}));
-  };
-
   const renderHeader = (item) => (
     <View style={styles.header}>
       <View style={styles.headerTextContainer}>
@@ -62,9 +42,8 @@ const SubTypeAccordion = ({
       </View>
       <View style={{ flexDirection: 'row' }}>
         <MarkedIcon
-          count={markedResourceCount}
-          onClick={handleMarkedClick}
-          hasMarked={markedResourceCount > 0}
+          resourceIds={resourceIds}
+          showCount
         />
         <CountIcon
           shape="square"
@@ -112,28 +91,18 @@ SubTypeAccordion.propTypes = {
   resourceIds: arrayOf(string.isRequired).isRequired,
   selectedCollectionId: string.isRequired,
   subType: string.isRequired,
-  updateMarkedResources: func.isRequired,
-  markedResources: shape({
-    marked: shape({}).isRequired,
-    lastMarked: shape({}).isRequired,
-  }).isRequired,
 };
 
 const mapStateToProps = (state) => ({
   selectedCollectionId: state.selectedCollection,
-  markedResources: markedResourcesSelector(state),
 });
 
 const mapDispatchToProps = {
   addResourceToCollectionAction: addResourceToCollection,
   removeResourceFromCollectionAction: removeResourceFromCollection,
-  updateMarkedResources: (resourceIdsMap) => ({
-    type: actionTypes.UPDATE_MARKED_RESOURCES,
-    payload: resourceIdsMap,
-  }),
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(React.memo(SubTypeAccordion));
+export default connect(mapStateToProps, mapDispatchToProps)(SubTypeAccordion);
 
 const styles = StyleSheet.create({
   header: {
